requirements.txt
# Core dependencies
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.19.0
pillow>=8.0.0

# Data processing
pandas>=1.3.0
scikit-learn>=0.24.0
scikit-image>=0.18.0

# Visualization
matplotlib>=3.3.0
seaborn>=0.11.0

# Configuration
pyyaml>=5.4.0

# Progress tracking
tqdm>=4.60.0

# Domain generalization (optional, for DomainBed integration)
wilds>=2.0.0

# Medical imaging (optional)
opencv-python>=4.5.0
albumentations>=1.0.0

# Experiment tracking (optional)
tensorboard>=2.6.0
wandb>=0.12.0

# Testing
pytest>=6.2.0
pytest-cov>=2.12.0

